requirements.txt
PyGithub==2.1.1
google-generativeai==0.3.2
python-dotenv==1.0.0
GitPython==3.1.40
